Integrate scoped alias regressions into existing tests
